requirements.txt
langchain 
streamlit-chat
fastembed
openai
langchainhub
chromadb
streamlit
watchdog
bs4
tiktoken
sentence-transformers
python-dotenv
docx2txt
unstructured
faster-whisper